Region of interest selection. Miss Calibration Detection

+++ get_actigraph_samplerate.m
@@ -7,5 +7,5 @@ fclose(fid);
 
 hz_string = regexp(line, '[0-9]+ Hz', 'match');
 scan_result = textscan(hz_string{1}, '%d');
-samplerate = scan_result{1};
+samplerate = double(scan_result{1});
 end
